+++ mercury/mercury-library/agent-didcommx/src/main/scala/io/iohk/atala/mercury/AgentService.scala
@@ -53,11 +53,11 @@ object AgentService {
     )
   )
 
-  val charlie = ZLayer.succeed(
-    AgentService[Agent.Charlie.type](
-      new DIDComm(UniversalDidResolver, CharlieSecretResolver.secretResolver),
-      Agent.Charlie
-    )
-  )
+  // val charlie = ZLayer.succeed(
+  //   AgentService[Agent.Charlie.type](
+  //     new DIDComm(UniversalDidResolver, CharlieSecretResolver.secretResolver),
+  //     Agent.Charlie
+  //   )
+  // )
 
 }