+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/model/IssueCredentialRecord.scala
@@ -6,8 +6,11 @@ import io.iohk.atala.mercury.protocol.issuecredential.OfferCredential
 import io.iohk.atala.mercury.protocol.issuecredential.RequestCredential
 import io.iohk.atala.mercury.protocol.issuecredential.IssueCredential
 import IssueCredentialRecord._
+import java.time.Instant
 final case class IssueCredentialRecord(
     id: UUID,
+    createdAt: Instant,
+    updatedAt: Option[Instant],
     thid: UUID,
     schemaId: Option[String],
     role: Role,